jwt_backend/pyproject.toml
[tool.poetry]
name = "jwt-backend"
version = "0.1.0"
description = ""
authors = ["Конаков Павел <konakov.po@yandex.ru>"]
readme = "README.md"
packages = [{include = "jwt_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
sqlmodel = "^0.0.8"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
uvicorn = "^0.21.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
